Prod-0: Move factories from factory package to domain package
